% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{accessibility_DA_location}
\alias{accessibility_DA_location}
\title{Count of point data per DA}
\usage{
accessibility_DA_location(DA_table)
}
\arguments{
\item{DA_table}{<\verb{sf data.frame}> A data.frame listing most recent DA ID with
spatial features.}
}
\value{
A list of two element. First is the count of each accessibility variable
for every DA. Second is a dictionary, with the industry title and explanation
for every variable code.
}
\description{
The function reads DMTI data from the AWS bucket and classifies each point
using sic code. It then sums each variable for every DA.
}
