% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{accessibility_mv_points}
\alias{accessibility_mv_points}
\title{Upload mapbox vector tileset for access points}
\usage{
accessibility_mv_points(points, username, access_token)
}
\arguments{
\item{points}{<\verb{named list}> Where the name is the variable (e.g. retail_general)
and the content is a data.frame of points.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
Uploads all the list to a mapbox vector tileset with "access_" to the
ID (access_retail_general).
}
\description{
Upload mapbox vector tileset for access points
}
