% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accessibility_point_dict}
\alias{accessibility_point_dict}
\title{Dictionary for pre-processed accessibility variables}
\format{
\subsection{\code{accessibility_point_dict}}{

A data frame of 3 columns
\describe{
\item{var_code}{The code used to refer to the variables}
\item{industry}{Name of the industry attached to the variable code}
\item{exp}{Explanation of the industry attached to the variable code}
}
}
}
\source{
Underlying data coming from DMTI. Industry name and explanation
retrieved using SIC codes come from
\url{https://www.naics.com/sic-industry-description}
}
\usage{
accessibility_point_dict
}
\description{
Computed using the \code{accessibility_DA_location()} function.
Industry name and explanation comes from the SIC code attached to every
point.
}
\keyword{datasets}
