% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{accessibility_polygons_helper}
\alias{accessibility_polygons_helper}
\title{Accessibility Polygons Helper Function}
\usage{
accessibility_polygons_helper(polygons)
}
\arguments{
\item{polygons}{<\code{sf}> An object of class \code{sf} containing the input
polygons to be processed.}
}
\value{
An object of class \code{sf} with the processed polygons, where
touching or intersecting polygons are merged into a single POLYGON object.
}
\description{
This function processes input polygons and groups them if they are touching
or intersecting. The grouped polygons are then returned as a single
POLYGON object.
}
