% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accessibility_themes}
\alias{accessibility_themes}
\title{All pre-processed accessibility themes}
\format{
\subsection{\code{accessibility_themes}}{

A character vector
}
}
\usage{
accessibility_themes
}
\description{
All pre-processed accessibility themes
}
\keyword{datasets}
