% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_rest_services.R
\name{arcgis_rest_services_ret}
\alias{arcgis_rest_services_ret}
\title{Retrieve GeoJSON Data from ArcGIS REST Services}
\usage{
arcgis_rest_services_ret(url)
}
\arguments{
\item{url}{<\code{character}> The base URL of the ArcGIS REST service endpoint.
The function ensures that the URL ends with "/query". If the URL does
not end with "/query", the function modifies it accordingly.}
}
\value{
An object of class \code{sf} representing the spatial dataframe containing
the fetched data.
}
\description{
This function makes a GET request to an ArcGIS REST service and retrieves
data in GeoJSON format. It is designed to fetch all features and fields from
the specified service endpoint. The retrieved data is returned as a spatial
dataframe using the \code{sf} package.
}
