% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_UCB.R
\name{BC_UCB}
\alias{BC_UCB}
\title{Retrieve Urban Containment Boundaries (UCB) for British Columbia Regions}
\usage{
BC_UCB(crs = 32610)
}
\value{
A list containing UCB data for the Capital Region District (\code{crd}),
Metro Vancouver (\code{metro_van}), and Nanaimo (\code{nanaimo}).
}
\description{
This function downloads the Urban Containment Boundaries (UCB) for various
regions in the province of British Columbia.
}
