% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{boc_interest_rates}
\alias{boc_interest_rates}
\title{Process Bank of Canada Interest Rate Data}
\usage{
boc_interest_rates()
}
\value{
A formatted data frame with interest rates from 1991 onwards, structured in a wide format.
}
\description{
This function downloads and processes Bank of Canada interest rate data
from the official Valet API. It extracts, reformats, and reshapes the data into a wide format,
keeping only observations from 1991 onwards.
}
