% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{bucket_list_content}
\alias{bucket_list_content}
\title{List bucket contents}
\usage{
bucket_list_content(bucket)
}
\arguments{
\item{bucket}{<\code{character}> The name of the bucket from which to download
the object.}
}
\value{
Returns a df of objects in the bucket
}
\description{
List bucket contents
}
