% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{bucket_read_object}
\alias{bucket_read_object}
\title{Read single object from AWS bucket}
\usage{
bucket_read_object(object, objectext, bucket, method)
}
\arguments{
\item{object}{<\code{character}> Name of the file with extension to retrieve.}

\item{objectext}{<\code{character}> Extension of the file to retrieve, e.g. \code{.csv}
or \code{.qs}.}

\item{bucket}{<\code{character}> The name of the bucket from which to download
the object.}

\item{method}{<\code{function}> A function used to read the object, e.g. if a \code{.csv}
then the method would be \code{utils::read.csv} or if a \code{.qs} would be \code{qs::qread}}
}
\value{
Returns the object read from the AWS bucket, and read using the
supplied \code{method}.
}
\description{
Read single object from AWS bucket
}
