% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{bucket_read_object_zip_shp}
\alias{bucket_read_object_zip_shp}
\title{Read a shapefile (.shp) from a zip file in an AWS S3 bucket}
\usage{
bucket_read_object_zip_shp(object, bucket)
}
\arguments{
\item{object}{<\code{character}> Name of the file with extension to retrieve.}

\item{bucket}{<\code{character}> The name of the bucket from which to download
the object.}
}
\value{
An \code{sf} object representing the shapefile.
}
\description{
This function downloads a zip file from an AWS S3 bucket, unzips it, and reads
the shapefile (.shp) contained in it. The function requires the \code{aws.s3}
package and the \code{sf} package.
}
\details{
This function assumes that the zip file in the bucket contains only
one shapefile. If there are multiple shapefiles, it will only read the first
one it finds.
}
