% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{bucket_write_folder}
\alias{bucket_write_folder}
\title{Write a folder to an AWS bucket}
\usage{
bucket_write_folder(folder, bucket, prune = "ask")
}
\arguments{
\item{folder}{<\code{character}> The folder that should be entirely uploaded to
the bucket.}

\item{bucket}{<\code{character}> The name of the bucket in which the entire folder
should be uploaded.}

\item{prune}{<\code{character}|\code{logical}> Should local files not present in the
remote bucket be deleted? Defaults to "ask", which will prompt the user to
confirm in the case that files will be deleted, but can also be set to \code{TRUE}
or \code{FALSE} to respectively delete or not delete local files without a prompt.}
}
\value{
An error message or nothing if ran succesfully.
}
\description{
Write a whole folder and its sub-folder to a pre-existing AWS bucket.
Environment variables (ID & KEY) to modify the buckets must be set beforehand, and
the user must have write access. As write requests are 10x more expensive than
read requests, each file is read from the bucket beforehand (if it exists) and if no
modification has been made to the file, it does not get written to the bucket.
}
