% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canale.R
\name{build_alp}
\alias{build_alp}
\title{Build ALP index for a given year}
\usage{
build_alp(
  years = cc.data::census_years[2:length(cc.data::census_years)],
  DA_table
)
}
\arguments{
\item{years}{<\code{numeric}> Indicating the years of the streets network to use.
Options are census years starting from 2001. Defaults to \code{\link{census_years}}.}

\item{DA_table}{<\verb{sf data.frame}> An \code{sf} object representing the
dissemination areas (DAs) to calculate the CanALE index for.}
}
\value{
A data.frame object representing the CanALE index for the given year
and DAs.
}
\description{
This function builds the Active Living Potential dataset
for a given year based on three measures: intersection density, dwelling
density, and points of interest (POI) density.
}
\details{
Workflow from CanALE User Guide http://canue.ca/wp-content/uploads/2018/03/CanALE_UserGuide.pdf
}
