% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canale.R
\name{build_canale}
\alias{build_canale}
\title{Build CanALE index for a given year}
\usage{
build_canale(year, DA_table)
}
\arguments{
\item{year}{<\code{numeric}> indicating the year of the streets network to use.
Options are census years starting from 2001.}

\item{DA_table}{<\verb{sf data.frame}> An \code{sf} object representing the
dissemination areas (DAs) to calculate the CanALE index for.}
}
\value{
A data.frame object representing the CanALE index for the given year
and DAs.
}
\description{
This function builds the Canadian Active Living Environment (CanALE) index
for a given year based on three measures: intersection density, dwelling
density, and points of interest (POI) density.
}
