% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defavorisation.R
\name{build_defavorisation}
\alias{build_defavorisation}
\title{Build Defavorisation Index for Given Years}
\usage{
build_defavorisation(years = c("2021", "2016", "2011"))
}
\arguments{
\item{years}{<\verb{character vector}> A vector of years to retrieve the defavorisation data for.
Defaults to 2021, 2016, and 2011.}
}
\value{
A \code{tibble} with the defavorisation indices for each year.
Columns include the DA identifier (\code{ID}), the material deprivation index for
each year (\code{defav_material_YEAR}), and the social deprivation index for each
year (\code{defav_social_YEAR}).
}
\description{
This function builds the defavorisation dataset for a given set of years.
It processes shapefiles containing material and social deprivation indices
for Quebec and interpolates the indices to match a new dissemination area (DA) structure.
}
\details{
The defavorisation dataset contains material and social deprivation
indices (QuintMatRS and QuintSocRS). The function interpolates previous years'
data to match the most recent DA structure and processes missing values.
The data used is sourced from the official Quebec deprivation indices provided by
Données Québec.
}
