% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_mixity.R
\name{build_social_mixity}
\alias{build_social_mixity}
\title{Build social mixity index}
\usage{
build_social_mixity(
  DB_table = bucket_read_object_zip_shp("DB_shp_carto.zip", "curbcut.rawdata")
)
}
\arguments{
\item{DB_table}{<\verb{sf data.frame}> An \code{sf} object representing the
dissemination areas (DAs) to calculate the CanALE index for.}
}
\value{
A data.frame object representing the CanALE index for the given year
and DAs.
}
\description{
This function builds the Social Mixity index dataset
}
