% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buildings_osm_ms_keys}
\alias{buildings_osm_ms_keys}
\title{Links and key from where to download the buildings dataset}
\format{
\subsection{\code{buildings_osm_ms_keys}}{

A data frame of 2 columns and the same number of 13 rows
\describe{
\item{osm_link }{The link from where to download buildings data from OSM}
\item{canbics_2021}{The province key from which to download MS building data}
\item{addresses}{The download link to get the province's database of addresses}
}
}
}
\source{
\url{https://github.com/microsoft/CanadianBuildingFootprints}
}
\usage{
buildings_osm_ms_keys
}
\description{
Links and key from where to download the buildings dataset
}
\keyword{datasets}
