% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings_sf.R
\name{buildings_sf}
\alias{buildings_sf}
\title{Build buildings and place each Province's processed data in a folder}
\usage{
buildings_sf(DA_processed_table, dest_folder, OSM_cache = TRUE)
}
\arguments{
\item{DA_processed_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame from
which the DA ID will be attached to every single building.}

\item{dest_folder}{<\code{character}> Destination folder where each processed
building dataset will be saved.}

\item{OSM_cache}{<\code{logical}> Should the OSM building data be downloaded
and processed with the \code{osmextract} package, or should it be read from the
cache. The cache needs to be set as a variable environment
\code{OSMEXT_DOWNLOAD_DIRECTORY}.}
}
\value{
Returns an error or nothing if ran successfully. The buildings of each
province are saved in the destination folder.
}
\description{
\code{buildings_sf()} uses a downloads building data from
Open Street Map and Microsoft's Canadian Building Footprints, processed is
and save each province's dataset in a destination folder.
}
