% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{canale_process}
\alias{canale_process}
\title{Process raw Can-ALE data from the AWS bucket}
\usage{
canale_process()
}
\value{
Returns the Can-ALE dataset processed country-wide.
}
\description{
Process raw Can-ALE data from the AWS bucket
}
