% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canbics_process.R
\name{canbics_process}
\alias{canbics_process}
\title{Process raw Canbics data from the AWS bucket}
\usage{
canbics_process()
}
\value{
Returns the Canbics dataset processed country-wide.
}
\description{
Process raw Canbics data from the AWS bucket
}
