% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{cansim_income_inequity}
\alias{cansim_income_inequity}
\title{Calculer un indicateur d'inégalité basé sur les revenus extrêmes}
\usage{
cansim_income_inequity()
}
\value{
Un data.frame au format large avec \code{id} (GeoUID) et les colonnes \code{income_ratio_ineq_YYYY}.
}
\description{
Cette fonction calcule le ratio entre le nombre de personnes avec un revenu de 250 000 $ ou plus
et celles avec un revenu inférieur à 50 000 $, pour une sélection de CMA canadiennes.
Les deux parties d'Ottawa-Gatineau sont agrégées.
}
