% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{cansim_income_median}
\alias{cansim_income_median}
\title{Extraire les revenus médians annuels pour les CMA du Canada}
\usage{
cansim_income_median()
}
\value{
Un data.frame au format large avec un identifiant \code{id} (GeoUID) par CMA, et une colonne par année.
}
\description{
Cette fonction télécharge les données de revenu total médian des personnes ayant un revenu
(table StatCan 11-10-0008-01), filtre pour les CMA sélectionnées, fait la correspondance
avec une table de noms géographiques, puis retourne un tableau en format wide avec une colonne
\code{id} (GeoUID) et des colonnes \code{income_median_ann_YYYY} pour chaque année.
}
