% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{cansim_unemployment}
\alias{cansim_unemployment}
\title{Process Unemployment Data from CANSIM}
\usage{
cansim_unemployment()
}
\value{
A data frame with one row per geographic unit and columns for unemployment rates by month (YYYYMM).
}
\description{
Downloads and reshapes CANSIM table 14-10-0287-03 to get monthly unemployment rates
for each geography (excluding Canada) in wide format (one column per month).
}
