% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{cansim_wages}
\alias{cansim_wages}
\title{Process Wage Rate Index Data from CANSIM}
\usage{
cansim_wages()
}
\arguments{
\item{table_id}{A string specifying the CANSIM table ID (default: "18-10-0140-01").}
}
\value{
A list of two data frames:
\item{union_wage_index_basic}{Basic union wage rate indexes by CMA}
\item{union_wage_index_supp}{Union wage rate indexes with supplements by CMA}
}
\description{
Retrieves and processes CANSIM wage rate index data, returning two datasets:
one for basic construction union wage rates and one for those including selected pay supplements.
}
