% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_utils.R
\name{census_add_parent_vectors}
\alias{census_add_parent_vectors}
\title{Add parent vectors to a set of census vectors}
\usage{
census_add_parent_vectors(census_vectors)
}
\arguments{
\item{census_vectors}{<\code{character}> One or more vectors from
\code{\link[cc.data]{census_vectors}}}
}
\value{
A character vector with the same elements as \code{census_vectors},
but also including the parent vectors of those elements, if any.
}
\description{
Given a set of census vectors, this function retrieves the corresponding
parent vectors from the census vectors table and adds them to the output.
}
