% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_agg_type.R
\name{census_agg_type}
\alias{census_agg_type}
\title{Get census variables aggregation type}
\usage{
census_agg_type(census_vectors, census_scales, census_years)
}
\arguments{
\item{census_vectors}{<\code{data.frame}> Should be equal to
\code{\link[susdata]{census_vectors}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[susdata]{census_years}}}
}
\value{
A named list of length two. The first is a vector of character of
additive variables, and the other is average.
}
\description{
Get the aggregation type which will be used for interpolation.
}
