% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_all_ids}
\alias{census_all_ids}
\title{All census ID of each scale, from the latest census}
\format{
\subsection{\code{census_all_ids}}{

A named list the same length as the length of \code{cc.data::census_scales}
}
}
\usage{
census_all_ids
}
\description{
All census ID of each scale, from the latest census
}
\keyword{datasets}
