% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_DA_years_dictionary.R
\name{census_build_DA_years_dictionary}
\alias{census_build_DA_years_dictionary}
\title{Create the dictionary linking the most recent IDs to past DA IDs}
\usage{
census_build_DA_years_dictionary(
  DA_data_raw,
  census_years = cc.data::census_years
)
}
\arguments{
\item{DA_data_raw}{<\verb{list of sf data.frame}> The DA output of
\code{\link[cc.data]{census_data_raw}}. The DA output of the raw data retrieval.}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[cc.data]{census_years}}}
}
\value{
A tibble with the number of columns as there are available years in the
census. The anchor is the most recent census' DA IDs. Every other column is
the other year's IDs with which the most recent DAs intersect.
}
\description{
Create the dictionary linking the most recent IDs to past DA IDs
}
