% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_custom_boundaries.R
\name{census_custom_boundaries}
\alias{census_custom_boundaries}
\title{Interpolate and process data for non-census boundaries}
\usage{
census_custom_boundaries(
  destination,
  DA_IDs,
  census_vectors = cc.data::census_vectors,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years,
  agg_type = census_agg_type(census_vectors = census_vectors, census_scales =
    census_scales, census_years = census_years),
  crs
)
}
\arguments{
\item{destination}{<\verb{named list}> A named list of sf data.frame. The name is
the name of a scale which can be found in the census, e.g. \code{CSD}, \code{CT}, \code{DA}.
Each contains an sf data.frame of non-census boundaries, for which data must
be interpolated. e.g. boroughs in Montreal.}

\item{DA_IDs}{<\verb{named list}> <\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database
to do interpolation.}

\item{census_vectors}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_vectors}}, or a subset of the character vector to
only filter the wanted variables.}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[cc.data]{census_years}}}

\item{agg_type}{<\verb{named list}> The output of
\code{\link[cc.data]{census_agg_type}}}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
The same list as what is fed to the \code{destination} argument,
with all processed census data appended.
}
\description{
Interpolate and process data for non-census boundaries
}
