% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_DA_years_dict}
\alias{census_DA_years_dict}
\title{DA ID years dictionary}
\format{
\subsection{\code{census_DA_years_dict}}{

A data.frame of the same number of columns as there are years available in
the census
\describe{
\item{ID_20xx}{All DA IDs of the most recent census}
\item{ID_...}{List of that year's DA IDs for which the DA ID of the most
recent census years intersects.}
}
}
}
\usage{
census_DA_years_dict
}
\description{
Dictionary to map all IDs of the most recent census to the DA IDs of the
previous censuses.
}
\keyword{datasets}
