% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_raw.R
\name{census_data_raw}
\alias{census_data_raw}
\title{Download census vectors and join to geometries for specific variables.}
\usage{
census_data_raw(
  empty_geometries,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years,
  var_codes = NULL
)
}
\arguments{
\item{empty_geometries}{list [\link{scale}][\link{year}] sf(ID, geometry) in target CRS}

\item{census_scales}{character, e.g., c("C","PR","CMA","CSD","CT","DA")}

\item{census_years}{numeric, e.g., c(1996, 2001, ..., 2021)}

\item{var_codes}{character|NULL subset of var_code; NULL = all (+parents)}
}
\value{
list[\link{scale}][\link{year}] sf(ID, geometry, \if{html}{\out{<vars>}})
}
\description{
Download census vectors and join to geometries for specific variables.
}
