% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_full_geos.R
\name{census_full_geos}
\alias{census_full_geos}
\title{Append Geospatial Data to Processed Census Data}
\usage{
census_full_geos(processed_census)
}
\arguments{
\item{processed_census}{<\verb{named list}> A list of data frames, each containing
census data at a different scale.}
}
\value{
A list of spatial data frames, each containing the
merged census and geospatial data for a particular scale.
}
\description{
This function retrieves the geospatial data for each scale in the
processed census data from a specified bucket, merges this data
with the existing census data, and returns the combined data set.
}
