% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_utils.R
\name{census_get_vectors_table}
\alias{census_get_vectors_table}
\title{Get census vectors table for a given set of vectors}
\usage{
census_get_vectors_table(census_vectors)
}
\arguments{
\item{census_vectors}{<\code{character}> One or more vectors from
\code{\link[cc.data]{census_vectors}}}
}
\value{
A data frame with the same columns as the census vectors table, but
filtered to include only the variables specified in \code{census_vectors} and
their parent vectors.
}
\description{
Given a set of census vectors, this function retrieves the corresponding
entries and the parent entries in the census vectors table. The output is a
data frame with one row per variable and several columns with information
about the variable.
}
