% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_interpolate.R
\name{census_interpolate}
\alias{census_interpolate}
\title{Interpolate all census years to the current year's geography}
\usage{
census_interpolate(
  data_raw,
  census_vectors = cc.data::census_vectors,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years,
  agg_type = NULL
)
}
\arguments{
\item{data_raw}{\if{html}{\out{<list of sf data.frame>}} The output of cc.data::census_data_raw.}

\item{census_vectors}{\if{html}{\out{<character vector>}} Should be equal to cc.data::census_vectors}

\item{census_scales}{\if{html}{\out{<character vector>}} Should be equal to cc.data::census_scales}

\item{census_years}{\if{html}{\out{<numeric vector>}}   Should be equal to cc.data::census_years}

\item{agg_type}{<named list|NULL> If NULL, built from variables actually present in data_raw.}
}
\value{
list [\link{scale}][\link{year}] sf with geometries of max year and values interpolated.
}
\description{
Interpolate all census years to the current year's geography
}
