% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_normalize.R
\name{census_normalize}
\alias{census_normalize}
\title{Normalize percentage variables}
\usage{
census_normalize(
  interpolated,
  census_vectors_table = cc.data::census_vectors_table,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years,
  unit_type = census_unit_type(census_vectors_table = census_vectors_table, census_scales
    = census_scales, census_years = census_years)
)
}
\arguments{
\item{interpolated}{<\verb{list of sf data.frame}> The output of
\code{\link[cc.data]{census_interpolate}}.}

\item{census_vectors_table}{<\code{data.frame}> Should be equal to
\code{\link[cc.data]{census_vectors_table}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[cc.data]{census_years}}}

\item{unit_type}{<\code{data.frame}> The output of
\code{\link[cc.data]{census_unit_type}}.}
}
\value{
A list of scales and years of census data with variables desired as
percentage normalized.
}
\description{
Normalize percentage variables
}
