% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_reduce_years.R
\name{census_reduce_years}
\alias{census_reduce_years}
\title{Reduce all census years to one dataframe}
\usage{
census_reduce_years(
  parent_dropped,
  census_vectors = cc.data::census_vectors,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years
)
}
\arguments{
\item{parent_dropped}{\if{html}{\out{<list of sf data.frame>}} The output of \code{census_drop_parents()}.}

\item{census_vectors}{\if{html}{\out{<character vector>}} Should be equal to \code{census_vectors}.}

\item{census_scales}{\if{html}{\out{<character vector>}} Should be equal to \code{census_scales}.}

\item{census_years}{\if{html}{\out{<numeric vector>}}   Should be equal to \code{census_years}.}
}
\value{
A list of scales with years reduced, and the years appended to all
variables column name
}
\description{
Reduce all census years to one dataframe
}
\seealso{
\code{census_drop_parents()}
}
