% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_scales}
\alias{census_scales}
\title{All available census scales}
\format{
\subsection{\code{census_scales}}{

A character (year) vector
}
}
\usage{
census_scales
}
\description{
All available census scales
}
\keyword{datasets}
