% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_full_geos.R
\name{census_switch_full_geo}
\alias{census_switch_full_geo}
\title{Retrieve and merge full census geospatial data with census df}
\usage{
census_switch_full_geo(df, scale_name)
}
\arguments{
\item{df}{<\code{data.frame}> The data frame containing the census data.}

\item{scale_name}{<\code{character}> The name of the scale used to identify
the corresponding census data ("CSD", "CT", ...)..}
}
\value{
A spatial data frame containing the merged census and geospatial
data for a particular scale.
}
\description{
This function reads the geospatial data from a specified bucket and
merges it with the given data frame of census information.
The geospatial data is identified by the scale name, and it is returned
as a spatial data frame.
}
