% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_units_type.R
\name{census_unit_type}
\alias{census_unit_type}
\title{Get census variables units type}
\usage{
census_unit_type(census_vectors, census_scales, census_years)
}
\arguments{
\item{census_vectors}{<\code{data.frame}> Should be equal to
\code{\link[susdata]{census_vectors}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[susdata]{census_years}}}
}
\value{
A dataframe of two columns. The first column is the var_code, and the
second is the unit.
}
\description{
Get census variables units type
}
