% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_vectors}
\alias{census_vectors}
\title{All census variable codes in a vector}
\format{
\subsection{\code{census_vectors}}{

A character vector
}
}
\usage{
census_vectors
}
\description{
All census variable codes in a vector
}
\keyword{datasets}
