% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_vectors_details}
\alias{census_vectors_details}
\title{All census vectors details used for data export on Curbcut}
\format{
\subsection{\code{census_vectors_details}}{

A data.frame the same number of rows as there are Curbcut
\code{\link[cc.data]{census_vectors}}
\describe{
\item{var_code}{The code used to refer to the variables}
\item{vec}{The \code{cancensus} code for the vector}
\item{vec_label}{The name of the variable}
\item{parent_vec}{The \code{cancensus} code for the parent vector}
\item{aggregation}{Indicating whether the value is additive or a
transformation}
\item{parent_vec_label}{The name of the parent variable}
}
}
}
\usage{
census_vectors_details
}
\description{
All census vectors details used for data export on Curbcut
}
\keyword{datasets}
