% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_vectors_education}
\alias{census_vectors_education}
\title{Education census vectors}
\format{
\subsection{\code{census_vectors_education}}{

A data.frame of census variables code
\describe{
\item{var_code}{See \code{\link[susbuildr]{add_variable}}}
\item{type}{See \code{\link[susbuildr]{add_variable}}}
\item{theme}{See \code{\link[susbuildr]{add_variable}}}
\item{vec_...}{The \code{cancensus} code for the vector at a particular year}
\item{parent_vectors}{The \code{cancensus} code for parent vectors}
\item{var_title}{See \code{\link[susbuildr]{add_variable}}}
\item{var_short}{See \code{\link[susbuildr]{add_variable}}}
\item{explanation}{See \code{\link[susbuildr]{add_variable}}}
}
}
}
\usage{
census_vectors_education
}
\description{
The variable code for each year
}
\keyword{datasets}
