% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_years}
\alias{census_years}
\title{All available census years}
\format{
\subsection{\code{census_years}}{

A numeric (year) vector
}
}
\usage{
census_years
}
\description{
All available census years
}
\keyword{datasets}
