% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cma_ct_ids}
\alias{cma_ct_ids}
\title{CT identifiers for all CMAs in Canada}
\format{
\subsection{\code{cma_ct_ids}}{

A tibble with two columns:
\describe{
\item{id}{Census tract identifier (GeoUID, 2021 geography).}
\item{cma_uid}{CMA identifier (CMA_UID).}
}
}
}
\usage{
cma_ct_ids
}
\description{
CT identifiers for all CMAs in Canada
}
\keyword{datasets}
