% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cma_ids}
\alias{cma_ids}
\title{All available CMAs identifiant}
\format{
\subsection{\code{cma_ids}}{

A character (cma_id) vector
}
}
\usage{
cma_ids
}
\description{
All available CMAs identifiant
}
\keyword{datasets}
