% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_data.R
\name{cmhc_abbreviations}
\alias{cmhc_abbreviations}
\title{Apply CMHC Data Abbreviations to Columns and Table Names}
\usage{
cmhc_abbreviations(cmhc_cma_vectors)
}
\arguments{
\item{cmhc_cma_vectors}{A list containing CMHC data frames.}
}
\value{
A modified list of data frames with standardized names.
}
\description{
This function replaces long column names and table names with standardized abbreviations
using the predefined named list \code{cmhc_data_abbreviations}.
}
