% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_update.R
\name{cmhc_all_scales}
\alias{cmhc_all_scales}
\title{Geographic scales handled, in run order (cheapest first)}
\usage{
cmhc_all_scales()
}
\value{
Character vector \code{c("cma", "csd", "ct", "zone", "nbhd")}.
}
\description{
Geographic scales handled, in run order (cheapest first)
}
