% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_data.R
\name{cmhc_finalize_results}
\alias{cmhc_finalize_results}
\title{Finalize CMHC reshaped result}
\usage{
cmhc_finalize_results(reshaped_list)
}
\arguments{
\item{reshaped_list}{A list of reshaped data.frames.}
}
\value{
A cleaned list of data.frames.
}
\description{
Cleans up the final reshaped CMHC data frame:
\itemize{
\item Removes join artifacts (e.g., Year.x, Year.y)
\item Sorts columns by time
\item Keeps 'GeoUID' first
}
}
