% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_data.R
\name{cmhc_get_annual_survey_zone}
\alias{cmhc_get_annual_survey_zone}
\title{Retrieve Annual CMHC Data for Survey Zones}
\usage{
cmhc_get_annual_survey_zone(requests, survey_zones_ref)
}
\arguments{
\item{requests}{A list of request objects. Each must include the following elements:
\itemize{
\item \code{survey}: Character. The CMHC survey code (e.g., \code{"Scss"}).
\item \code{series}: Character. The data series (e.g., \code{"Starts"}).
\item \code{dimension}: Character. The breakdown dimension (e.g., \code{"Dwelling Type"}).
\item \code{years}: Integer vector. List of years to request (e.g., \code{2010:2024}).
}}

\item{survey_zones_ref}{A reference \code{data.frame} with columns \code{ZONE_NAME_EN} and \code{METZONE_UID}, used to attach unique \code{GeoUID} to each zone.}
}
\value{
A named list with element \code{survey_zone}, containing one cleaned \code{data.frame} per variable requested.
Each data.frame has one row per survey zone (\code{id}), and one column per year (e.g., \code{starts_dwelling_type_all_2010}, ...).
}
\description{
Downloads, cleans, reshapes, and merges annual CMHC data for survey zones associated with selected parent CMAs.
This function runs in parallel over all requested parent CMA GeoUIDs and returns data harmonized by zone and year.
}
