% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_data.R
\name{cmhc_get_arcgis_cma_uids}
\alias{cmhc_get_arcgis_cma_uids}
\title{Get CMA UIDs with CMHC Rental Market Survey coverage}
\usage{
cmhc_get_arcgis_cma_uids(vintage = "2021")
}
\arguments{
\item{vintage}{Character. Census vintage to filter against (default "2021").}
}
\value{
Sorted character vector of CMA UIDs (including cmasplit UIDs like "24505").
}
\description{
Queries CMHC's ArcGIS geospatial service (layer 4) to retrieve the set of
"centres" (CMAs + CAs) that have Rms survey data. UID is constructed as
\code{paste0(PRUID, sgc_cma_ca_cde)}. Results are filtered against our
geography.cma + geography.cmasplit tables to drop CAs we don't store.
}
