% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_data.R
\name{cmhc_get_arcgis_csd_uids}
\alias{cmhc_get_arcgis_csd_uids}
\title{Get CSD UIDs with CMHC Rental Market Survey coverage}
\usage{
cmhc_get_arcgis_csd_uids(vintage = "2021")
}
\arguments{
\item{vintage}{Character. Census vintage to filter against (default "2021").}
}
\value{
Sorted character vector of CSD UIDs.
}
\description{
Queries CMHC's ArcGIS geospatial service (layer 3) to retrieve the set of
Census Subdivisions that have Rms survey data, then filters against our
geography.csd table to keep only CSDs we store.
}
