% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_4square.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to Snowflake Database}
\usage{
connect_to_db()
}
\value{
A \code{DBI} connection object.
}
\description{
This function establishes a connection to the Snowflake database using credentials stored in \code{.Renviron}.
}
