% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_connect}
\alias{db_connect}
\title{Connect to the AWS MySQL database}
\usage{
db_connect()
}
\value{
returns an S4 object that inherits from \code{\link[DBI]{DBIConnection-class}}.
This object is used to communicate with the database engine.
}
\description{
Connect to the AWS MySQL database using variables saved in the .Renviron
}
