% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from the AWS MySQL database}
\usage{
db_disconnect(conn)
}
\arguments{
\item{conn}{A \code{\link[DBI]{DBIConnection-class}} object, as returned by
\code{\link[cc.data]{db_connect}}.}
}
\value{
A message out of \code{\link[DBI]{dbDisconnect}}
}
\description{
Disconnect from the AWS MySQL database
}
