% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_get.R
\name{DB_get}
\alias{DB_get}
\title{Retrieve and Process Dissemination Block Geographical Information}
\usage{
DB_get(years = cc.data::census_years, crs = NULL, DA_carto = NULL)
}
\arguments{
\item{years}{<\verb{numeric vector}> A vector of years to process. Default is set to
census_years from the cc.data package. The year 1996 is automatically excluded
as it does not contain DBs.}

\item{crs}{<\code{character}> Coordinate Reference System to be used for transforming
the spatial data.}

\item{DA_carto}{<\verb{sf object}> An sf object representing cartographic boundaries
for intersection with the DB geographical information.}
}
\value{
A list of processed spatial data for each year, each element being an sf
object with MULTIPOLYGON geometry, and containing population and dwelling counts.
The data is transformed to the specified CRS and intersected with the DA_carto object.
}
\description{
This function processes geographical information from Dissemination Blocks (DBs)
for a specified set of years. It excludes the year 1996 as there were no DBs in that
year, transforms, and intersects the data with given cartographic boundaries,
and appends population and dwelling data. It supports specific transformation
and intersection operations on spatial data.
}
