% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_read_all_table}
\alias{db_read_all_table}
\title{Read all data from a specified table}
\usage{
db_read_all_table(table)
}
\arguments{
\item{table}{<\code{character}> String indicating the name of the table to read}
}
\value{
A data frame containing all data from the specified table
}
\description{
Read all data from a specified table
}
