% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_read_data}
\alias{db_read_data}
\title{Read a table in the MySQL database}
\usage{
db_read_data(table, columns = "*", column_to_select = "ID", IDs, crs = 3347)
}
\arguments{
\item{table}{<\code{character}> The table name in the MySQL database. To list all
tables, use \code{\link[DBI]{dbListTables}}.}

\item{columns}{<\verb{character vector}> A character vector of all columns to
retrieve. By default, all columns (\code{*}).}

\item{column_to_select}{<\verb{character vector}> To column from which the \code{IDs}
should be selected. Defaults to \code{ID}.}

\item{IDs}{<\verb{character vector}> A character vector of all IDs to retrieve.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to which the data
should be transformed.}
}
\value{
A tibble or an sf tibble depending if geometry is available.
}
\description{
Read a table in the database by filtering the ID column.
}
