% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_read_data_helper}
\alias{db_read_data_helper}
\title{Helper function to read data from a database table based on selected columns
and IDs}
\usage{
db_read_data_helper(table, columns = "*", column_to_select = "ID", IDs = NULL)
}
\arguments{
\item{table}{<\code{character}> The table name in the MySQL database. To list all
tables, use \code{\link[DBI]{dbListTables}}.}

\item{columns}{<\verb{character vector}> A character vector of all columns to
retrieve. By default, all columns (\code{*}).}

\item{column_to_select}{<\verb{character vector}> To column from which the \code{IDs}
should be selected. Defaults to \code{ID}.}

\item{IDs}{<\verb{character vector}> A character vector of all IDs to retrieve.}
}
\value{
A tibble containing the retrieved data from the database table.
}
\description{
Helper function to read data from a database table based on selected columns
and IDs
}
\details{
If columns are specified, this function will retrieve the specified columns
and IDs. If "*" is used or the columns argument, all columns will be retrieved.

The function will split the IDs into multiple smaller calls if the length of
the IDs exceeds 2500.

The retrieved data is returned as a tibble. The function converts each column
to its right class based on its data type. ID columns are converted to
character and all other columns are converted using the utils::type.convert()
function.
}
