% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_read_long_table}
\alias{db_read_long_table}
\title{Read long table from the MySQL database}
\usage{
db_read_long_table(table, DA_ID)
}
\arguments{
\item{table}{<\code{character}> The table name in the MySQL database. To list all
tables, use \code{\link[DBI]{dbListTables}}.}

\item{DA_ID}{<\verb{vector of character}> DA IDs from which to retrieve all their
spatial features from the \code{table}.}
}
\value{
A tibble or an sf tibble depending if geometry is available.
}
\description{
Read data from a long table. The table must have been previously created
using \code{\link[cc.data]{db_write_long_table}}
}
