% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_read_ttm}
\alias{db_read_ttm}
\title{Read and return a travel time matrix from a database}
\usage{
db_read_ttm(mode, DA_ID)
}
\arguments{
\item{mode}{<\code{character}> The mode of transportation (e.g. "car", "bicycle",
"foot", "transit_pwd", ...)}

\item{DA_ID}{<\code{character}> A vector of DA IDs to retrieve data for}
}
\value{
A data frame containing the travel time matrix
}
\description{
Read and return a travel time matrix from a database
}
