% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_write_long_table}
\alias{db_write_long_table}
\title{Write long table to the MySQL database}
\usage{
db_write_long_table(df, tb_name)
}
\arguments{
\item{df}{<\code{data.frame}> A data.frame or an sf data.frame containing a large
number of rows, big enough that it will need its own DA dictionary.}

\item{tb_name}{<\code{character}> The name of the table to be in the database,
e.g. \code{buildings} or \code{street}.}
}
\value{
Returns an error or nothing if ran successfully. The \code{df} is saved
and the name of the table corresponds to \code{tb_name} in the database. And there
is another \code{tb_name_DA_dict} table which is DA dictionary. All ID of \code{df} are
regrouped per DA ID in a JSON column.
}
\description{
Write long table to the MySQL database
}
