% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_write_table_helper}
\alias{db_write_table_helper}
\title{Write data frame to database table}
\usage{
db_write_table_helper(
  df,
  tb_name,
  primary_key = NULL,
  index = NULL,
  rows_per_append_wave = 1000
)
}
\arguments{
\item{df}{<\code{data.frame}> A non-sf data.frame to be written to the MySQL}

\item{tb_name}{<\code{character}> The name of the table in the MySQL database}

\item{primary_key}{<\code{character}> A column of the table to be written that should
be the pimary key for faster retrieval. The ID in that column needs to be
unique.}

\item{index}{<\code{character}> A column of the table to be written that should
be indexed for faster retrieval. No need for the values of the column to be
unique.}

\item{rows_per_append_wave}{<\code{numeric}> Number of rows to each for each waves
of append. The thiner the dataframe is, the more many rows can be appended
at a time. For census data with lots of columns and geometry columns, as
each row is heavier, the number of rows to append at once needs to be smaller
(around 1000). For buildings, every row is smaller and so around 50k can
be appended at the same time.}
}
\description{
This function writes a data frame to a database table. If a primary key is
provided, the function creates a table with the primary key for faster
retrieval and appends the data by waves. If an index is provided, the function
creates an index on the index column for faster retrieval.
}
