% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_write_ttm}
\alias{db_write_ttm}
\title{Write travel time matrix to the MySQL database}
\usage{
db_write_ttm(ttm, mode)
}
\arguments{
\item{ttm}{<\verb{named list}> The travel time matrices in a list. The output
of \code{\link[cc.data]{tt_calculate}}.}

\item{mode}{<\code{character}> Mode to write in the title of each written dataframes,
e.g. \code{foot}, \code{bicycle}, \code{car}, \code{transit_pwd}, ...}
}
\value{
Returns nothing if successful. Every dataframe in the \code{ttm} list are
saved as a dataframe, with names such as \code{ttm_car_35210809}.
}
\description{
Write travel time matrix to the MySQL database
}
