% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.R
\name{forest_load_polygons}
\alias{forest_load_polygons}
\title{Load and merge CMM forest cover polygons}
\usage{
forest_load_polygons(shp_dir)
}
\arguments{
\item{shp_dir}{Character. Root directory where all unzipped forest cover
shapefiles are stored.}
}
\value{
An \code{sf} object with all forest cover polygons.
}
\description{
Given a directory containing all unzipped forest cover SHP folders/files
from the CMM, this function lists all .shp (case-insensitive), reads them
as sf, and merges them into a single sf object.
}
