% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{format_date_names}
\alias{format_date_names}
\title{Helper function: Reformat column names from "YYYY-MM" to "YYYYMM"}
\usage{
format_date_names(x)
}
\description{
Helper function: Reformat column names from "YYYY-MM" to "YYYYMM"
}
