% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canale.R
\name{get_all_three_plus_ways}
\alias{get_all_three_plus_ways}
\title{Get all three-plus way intersections in Canada}
\usage{
get_all_three_plus_ways(OSM_cache = TRUE)
}
\arguments{
\item{OSM_cache}{<\code{logical}> Indicates whether to use cached OSM data
or download it again. Default is TRUE.}
}
\value{
A data frame containing all the three-plus way intersections in
Canada.
}
\description{
This function returns all three-plus way intersections in Canada. It downloads
OpenStreetMap (OSM) data for Canada, retains footpaths, and excludes
limited-access highways and their entrances and exits. The OSM data is
transformed to the Stats Can CRS, and a grid is created to separate the work
charge to different cores. The function then calculates all the three-plus
way intersections in all the grids, and the streets that touch the lines of
the grid and potentially missed in the first step. Finally, it binds all the
three-plus way intersections and returns them.
}
