% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_raw.R
\name{get_census_data}
\alias{get_census_data}
\title{Get Census Data}
\usage{
get_census_data(
  census_dataset,
  var_codes,
  region,
  scale,
  cache = cancensus::list_cancensus_cache()
)
}
\arguments{
\item{census_dataset}{<\code{character}> A string representing the dataset identifier.}

\item{var_codes}{<\verb{character vector}> Containing variable codes to be retrieved,
named.}

\item{region}{<\verb{named list}> Specifying the region code (e.g., \code{list(C = "01")}).}

\item{scale}{<\code{character}> A string or numeric value representing the scale or
level of the data.}

\item{cache}{<\code{data.frame}> A data frame containing cached information, structured
appropriately. The output of \code{cancensus::list_cancensus_cache()}}
}
\value{
A data frame containing the requested census data. If the data is found
in the cache, it is returned directly from there. If not, the data is
retrieved vector by vector and merged by "GeoUID".
}
\description{
Retrieves census data for the given region and scale, utilizing caching if
available. If unavailable, iterate over every variable to make sure we
will be utilizing their cache next time.
}
