% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_gtfs.R
\name{gtfs_download_BC}
\alias{gtfs_download_BC}
\title{Download from GTFS (Only British Columbia)}
\usage{
gtfs_download_BC(dest_folder = tempdir())
}
\arguments{
\item{dest_folder}{Path to the directory to save the downloaded zip files.
If not provided, default is the system's temporary directory.}
}
\value{
Returns the \code{dest_folder}. Files are downloaded to the \code{dest_folder}
with a \code{gtfs_transit_feed} pattern in the file name.
}
\description{
Downloads GTFS data from https://database.mobilitydata.org/.
Filters for Canada and active zip files and downloads them to dest_folder.
}
