% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs_extract}
\alias{gtfs_extract}
\title{Extract GTFS to create a GTFS object}
\usage{
gtfs_extract(gtfs_zip = gtfs_combine(), maximum_walk_time = 800)
}
\arguments{
\item{gtfs_zip}{Path to the GTFS zip file to extract and create a transfer
table for. If not provided, the gtfs zip file is built using
\code{\link[cc.data]{gtfs_combine}}.}

\item{maximum_walk_time}{Maximum walking time in seconds for transfer connections.}
}
\value{
Returns a GTFS object.
}
\description{
Extract GTFS to create a GTFS object
}
