% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs_get_traveltimes}
\alias{gtfs_get_traveltimes}
\title{Get transit traveltimes from all transit stop to every transit stop.}
\usage{
gtfs_get_traveltimes(
  gtfs,
  day = "wednesday",
  start_time_limits = c(8, 8.5) * 3600,
  max_traveltime = 60 * 60
)
}
\arguments{
\item{gtfs}{GTFS object used for calculating travel times.
The output of \code{\link[cc.data]{gtfs_extract}}.}

\item{day}{Day of the week to calculate travel times for.
Valid options are "monday", "tuesday", "wednesday", "thursday", "friday",
"saturday", and "sunday".}

\item{start_time_limits}{Two-element numeric vector specifying the start
time range in seconds since midnight to calculate travel times for.}

\item{max_traveltime}{Maximum travel time in seconds.}
}
\value{
Named list of data frames, each containing the start time, duration,
and destination stop ID for each travel time.
}
\description{
Calculates travel times from a given GTFS zip file for a specific day
and time period. Returns a list of data frames, each containing the start
time, duration, and destination stop ID for each travel time.
}
