% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs_traveltime_matrix_final}
\alias{gtfs_traveltime_matrix_final}
\title{Get traveltime transit matrix}
\usage{
gtfs_traveltime_matrix_final(prep_output)
}
\arguments{
\item{prep_output}{<\code{list}> Output of
\code{\link[cc.data]{gtfs_traveltime_matrix_prep}}}
}
\value{
A named list containing the travel times from each origin point to
each destination point (both are IDs of \code{DA_table}).
}
\description{
PART 2 OF 2 FOR TRAVEL TIME MATRICE CALCULATIONS. PART 1:
\code{\link[cc.data]{gtfs_traveltime_matrix_prep}}
The two functions calculate travel times from a set of origin points to a set of
destination points using transit. It takes in a GTFS list, a traveltimes list,
and a table of origin/destination points (referred to as DA_table) and returns
a named list which, for every DA, lists every reachable DA and the travel time.
The user can also specify a maximum walk distance (default is 800 meters) and a
routing server for walking time from the origin/destination point to the transit
stops (default is on local machine using OSRM: "http://localhost:5000/". See
\code{\link[cc.data]{tt_local_osrm}}).
}
