% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs_traveltime_matrix_prep}
\alias{gtfs_traveltime_matrix_prep}
\title{Preparation to get traveltime transit matrix}
\usage{
gtfs_traveltime_matrix_prep(
  gtfs,
  traveltimes,
  DA_table,
  maximum_walk_dist = 800,
  routing_server = "http://localhost:5000/"
)
}
\arguments{
\item{gtfs}{<\code{gtfs}> A GTFS list containing information on transit. See
\code{\link[cc.data]{gtfs_combine}}}

\item{traveltimes}{<\verb{named list}> Contains information on travel times between
transit stops. See \code{\link[cc.data]{gtfs_get_traveltimes}}}

\item{DA_table}{<\verb{sf data.frame}> Object containing origin/destination
polygons. The origin/destination points will be the centroids.}

\item{maximum_walk_dist}{A numeric value specifying the maximum amount of
meters that a person is willing to walk to a transit stop.
Default is 800.}

\item{routing_server}{A string specifying the url of the routing server to
use for calculating walking times. Default is local machine using OSRM:
"http://localhost:5000/". See \code{\link[cc.data]{tt_local_osrm}})}
}
\value{
Final part \code{\link[cc.data]{gtfs_traveltime_matrix_final}}: A
named list containing the travel times from each origin point to
each destination point (both are IDs of \code{DA_table}).
}
\description{
PART 1 OF 2 FOR TRAVEL TIME MATRICE CALCULATIONS. PART 2:
\code{\link[cc.data]{gtfs_traveltime_matrix_final}}
The two functions calculate travel times from a set of origin points to a set of
destination points using transit. It takes in a GTFS list, a traveltimes list,
and a table of origin/destination points (referred to as DA_table) and returns
a named list which, for every DA, lists every reachable DA and the travel time.
The user can also specify a maximum walk distance (default is 800 meters) and a
routing server for walking time from the origin/destination point to the transit
stops (default is on local machine using OSRM: "http://localhost:5000/". See
\code{\link[cc.data]{tt_local_osrm}}).
}
