% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{list_accessibility_themes}
\alias{list_accessibility_themes}
\title{List pre-processed accessibility themes}
\usage{
list_accessibility_themes()
}
\value{
Returns a character vector of pre-processed accessibility themes
}
\description{
List pre-processed accessibility themes
}
