% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_census_vectors.R
\name{list_census_vectors}
\alias{list_census_vectors}
\title{List available census vectors}
\usage{
list_census_vectors()
}
\value{
Returns a data frame detailing the available Census vectors
(i.e. variables).
}
\description{
List available census vectors
}
