% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{lst_process}
\alias{lst_process}
\title{Process Land Surface Temperature Data Across Years}
\usage{
lst_process(DA_table)
}
\arguments{
\item{DA_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame that contains
the Dissemination Area (DA) information for spatial joins.}
}
\value{
A tibble containing the processed land surface data, including the merged
values across the specified years.
}
\description{
This function processes land surface temperature data across specified years by reading
the data using the \code{read_and_process_data_canue} function. It performs
a series of transformations and merges the data across different years
resulting in a tibble with the land surface values grouped by DA ID.
}
