% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi_features_to_point}
\alias{ndvi_features_to_point}
\title{Convert HLS Features to Point Data}
\usage{
ndvi_features_to_point(
  master_polygon,
  features,
  temp_folder = tempdir(),
  year,
  crs
)
}
\arguments{
\item{master_polygon}{<\code{sf}> Naster polygon as cartographic (DA_carto)}

\item{features}{<\code{data.frame}> A data.frame of \code{features} subsets retrieved
from collecetions. The output of \code{\link{ndvi_get_items_hls}}`. It must
include information about collections, granule IDs, cloud cover, bands, and
asset links.}

\item{temp_folder}{<\code{character}> Defaults to \code{tempdir()}. However, to save disk
space, it can be a ink to a managed temporary folder in a dedicated
directory. This allows to have complete control over permissions and other attributes
when paralleling. Temporary files are not removed otherwise, and the usual temporary
folder becomes to heavy.}

\item{year}{<\code{numeric}> Which years are the features from?}

\item{crs}{<\code{numeric}> CRS of the region.}
}
\value{
An \code{sf} object containing the NDVI points. The result includes the NDVI
values for each valid raster layer, and it is filtered to exclude poor quality
pixels.
}
\description{
This function takes a data frame of HLS features and processes them into point
data for further analysis. It includes downloading, cropping, and masking raster
data, along with calculating the Normalized Difference Vegetation Index (NDVI) and
quality filtering.
}
