% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi_get_features}
\alias{ndvi_get_features}
\title{Retrieve Harmonized Landsat Sentinel-2Features}
\usage{
ndvi_get_features(
  years = ndvi_years(),
  output_path = "calculated_ignore/ndvi/"
)
}
\arguments{
\item{years}{<\verb{numeric vector}> A vector of years for which NDVI features are
to be retrieved. Defaults to the result of \code{ndvi_years()}.}

\item{output_path}{<\code{character}> A character string specifying the path where the
retrieved data will be saved. Defaults to "calculated_ignore/ndvi/".}
}
\value{
NULL The function performs the retrieval and saving operations but
does not return any value.
}
\description{
This function retrieves Harmonized Landsat Sentinel-2 (HLS) features for
the specified years and saves them
in the specified output path. The data is collected from the CMR Earth Data
STAC API, with search criteria including HLS collection names, bounding box,
and growing season time range. The retrieved data is parallelized using
future_lapply and saved in the specified output path.
}
