% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi_get_items_hls}
\alias{ndvi_get_items_hls}
\title{Retrieve HLS Items for a Specific Year and Season}
\usage{
ndvi_get_items_hls(year, limit = 25, zone_bbox)
}
\arguments{
\item{year}{<\code{numeric}> The year for which the HLS items are to be retrieved.}

\item{limit}{<\code{numeric}> Maximum number of items to retrieve per page.
Default is 25.}

\item{zone_bbox}{<\verb{numeric vector}> Bounding box coordinates (xmin, ymin,
xmax, ymax) of the area to retrieve. EPSG4326}
}
\value{
A list of retrieved HLS items.
}
\description{
This function retrieves Harmonized Landsat Sentinel-2 (HLS) items
for a given year and limit within the growing season (from May 1st to August
31st). The function first performs an initial retrieval to determine the total
number of pages, and then utilizes parallel processing to fetch the required
data.
}
