% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi_import}
\alias{ndvi_import}
\title{Import NDVI Data for Specific Months and Process It}
\usage{
ndvi_import(
  years = ndvi_years(),
  output_path = "calculated_ignore/ndvi/",
  temp_folder = "calculated_ignore/ndvi/tmp/"
)
}
\arguments{
\item{years}{<\verb{numeric vector}> A vector of years for which NDVI features are
to be retrieved. Defaults to the result of \code{ndvi_years()}.}

\item{output_path}{<\code{character}> A character string specifying the path
where the calculated NDVI data should be saved. Defaults to "calculated_ignore/ndvi/".}

\item{temp_folder}{<\code{character}> Link to a managed temporary folder in a dedicated
directory. This allows to have complete control over permissions and other attributes
when paralleling. Temporary files are not removed otherwise, and the usual temporary
folder becomes to heavy.}
}
\value{
NDVI raters (.tiff) are saved in the output path.
}
\description{
This function imports NDVI data for specific months (May to August) and performs
various processing tasks. The function is intended to work with NDVI data for
analyzing vegetation patterns. It imports, filters, and processes the data
to be used in further analyses.
}
\details{
This function relies on several sub-functions to perform various tasks. The overall
functionality includes retrieving specific date ranges, importing NDVI data,
processing individual bands, calculating NDVI, and performing quality filtering.
}
