% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi_import_from_masterpolygon}
\alias{ndvi_import_from_masterpolygon}
\title{Import NDVI from Master Polygon}
\usage{
ndvi_import_from_masterpolygon(
  master_polygon,
  years = ndvi_years(),
  output_path,
  temp_folder = tempdir(),
  overwrite = FALSE,
  filter_cloudy_tiles = TRUE
)
}
\arguments{
\item{master_polygon}{<\verb{sf polygon}> The master polygon of the region.}

\item{years}{<\verb{character vector}> A vector specifying the years for which
NDVI data is to be retrieved. Defaults to the output of \code{ndvi_years()}.}

\item{temp_folder}{<\code{character}> A temporary folder path to store intermediate data.
Defaults to the system's temporary directory. However, to save disk
space, it can be a ink to a managed temporary folder in a dedicated
directory. This allows to have complete control over permissions and other attributes
when paralleling. Temporary files are not removed otherwise, and the usual temporary
folder becomes to heavy.}
}
\value{
NULL. The function processes the NDVI points and applies relevant
filtering and transformation but does not return a value.
}
\description{
This function reads the master polygon of a region and extracts NDVI data
for specified years. The process involves extracting the bounding box of the
master polygon, retrieving features related to the specified region, and
filtering the NDVI points based on tile information. It saves the yearly NDVI
values in the \code{output_path/year} with centroid of every 30m cells as the ID.
The satellite is very precise, and every raster is exactly the same centroid
value, which can be used as an ID. We keep geometry for the latest year,
and remove it from other years.
}
