% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{ndvi_process}
\alias{ndvi_process}
\title{Process raw NDVI data from the AWS bucket}
\usage{
ndvi_process(DA_table)
}
\arguments{
\item{DA_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame}
}
\value{
Returns the NDVI dataset processed country-wide.
}
\description{
Process raw NDVI data from the AWS bucket
}
