% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi_years}
\alias{ndvi_years}
\title{Generate a Sequence of Years from 2013 to Current Year}
\usage{
ndvi_years()
}
\value{
A numeric vector containing the sequence of years from 2013 to the current year.
}
\description{
This function creates a sequence of years from 2013 to the current year.
It is used for NDVI import, as the first satellite images were from 2013.
}
