% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{no2_process}
\alias{no2_process}
\title{Process raw NO2 data from the AWS bucket}
\usage{
no2_process(DA_table)
}
\arguments{
\item{DA_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame}
}
\value{
Returns the NO2 dataset processed country-wide.
}
\description{
Process raw NO2 data from the AWS bucket
}
