% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_places.R
\name{osm_strategy_pt_in_poly}
\alias{osm_strategy_pt_in_poly}
\title{Strategy: point-in-polygon deduplication}
\usage{
osm_strategy_pt_in_poly(d)
}
\arguments{
\item{d}{An \code{sf} data frame with a \code{layer_src} column from \code{\link[=osm_fetch_raw]{osm_fetch_raw()}}.}
}
\value{
A filtered \code{sf} data frame.
}
\description{
Keeps all polygons; drops points that fall inside one of those polygons
(because they refer to the same feature, tagged twice).
}
