% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_places.R
\name{osm_to_centroid}
\alias{osm_to_centroid}
\title{Collapse polygons to centroids}
\usage{
osm_to_centroid(x)
}
\arguments{
\item{x}{An \code{sf} data frame.}
}
\value{
An \code{sf} data frame with POINT geometry.
}
\description{
Ensures the resulting \code{sf} is pure POINT geometry (so it round-trips through
\code{upload_destination_pois()} and joins cleanly with other POI sources).
}
