% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_codes_process.R
\name{postal_codes_process}
\alias{postal_codes_process}
\title{Process raw postal codes from the AWS bucket}
\usage{
postal_codes_process(DA_processed_table)
}
\arguments{
\item{DA_processed_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame from
which the DA ID will be attached to every single building.}
}
\value{
An \code{sf} data.frame of 3 columns: postal code, dissemination area ID
and the centroid point geometry of every postal code in the country.
}
\description{
Process raw postal codes from the AWS bucket
}
