% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{process_constr_price}
\alias{process_constr_price}
\title{Process Construction Price Data from CANSIM}
\usage{
process_constr_price(table_id)
}
\arguments{
\item{table_id}{A string specifying the CANSIM table ID (default: "18-10-0276-01").}
}
\value{
A list containing three data frames:
\item{constr_price_apt}{Construction price data for apartment buildings}
\item{constr_price_resi}{Construction price data for residential buildings}
\item{constr_price_non_resi}{Construction price data for non-residential buildings}
}
\description{
This function retrieves construction price data from a specified CANSIM table, processes it,
and returns three cleaned datasets: one for apartment buildings, one for residential buildings,
and another for non-residential buildings. It also merges the data with geographic information.
}
