% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_4square.R
\name{process_poi_cma}
\alias{process_poi_cma}
\title{Extract and Process Points of Interest (POI) for Canadian CMAs}
\usage{
process_poi_cma(conn, cma_nash)
}
\arguments{
\item{conn}{A database connection object obtained from \code{connect_to_db()}.}

\item{cma_nash}{A spatial dataframe (\code{sf}) containing Canadian CMAs with geometries.}
}
\value{
An \code{sf} object containing POI evolution data per year for each CMA.
}
\description{
This function retrieves Points of Interest from a database, associates them with CMAs,
calculates the number of active POIs per year, and creates a structured dataset.
}
