% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time_matrix.R
\name{restart_osrm_container}
\alias{restart_osrm_container}
\title{Restart OSRM Docker container based on profile}
\usage{
restart_osrm_container(profile)
}
\arguments{
\item{profile}{Routing profile (bicycle, driving, foot)}
}
\value{
TRUE if successfully restarted, stops with error otherwise
}
\description{
Restart OSRM Docker container based on profile
}
