% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode_create_local_nominatim.R
\name{rev_geocode_create_local_nominatim}
\alias{rev_geocode_create_local_nominatim}
\title{Create a docker local Nominatim instance for faster reverse geocoding}
\usage{
rev_geocode_create_local_nominatim()
}
\value{
Opens a terminal from which the docker image is created.
}
\description{
Create a docker local Nominatim instance for faster reverse geocoding
}
