% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode.R
\name{rev_geocode_localhost}
\alias{rev_geocode_localhost}
\title{Reverse geocode using a local Nominatim instance (docker)}
\usage{
rev_geocode_localhost(point_sf, street = FALSE)
}
\arguments{
\item{point_sf}{< \code{sfc_POINT}> Point sf, e.g. a centroid of a building}

\item{street}{< \code{logical}> Is it to reverse geocode streets (No city name
appended)}
}
\value{
The address retrieved from the local Nominatim instance
}
\description{
Reverse geocode using a local Nominatim instance (docker)
}
