% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode_sf.R
\name{rev_geocode_sf}
\alias{rev_geocode_sf}
\title{Reverse geocode spatial features}
\usage{
rev_geocode_sf(prov_folder, nb_batches = 100, nb_sf_process = 400)
}
\arguments{
\item{prov_folder}{<\code{character}> Where the sf data.frames with spatial features
for each province live}
}
\value{
Override the existing tables in prov_folder with the addresses attached.
}
\description{
This geocoding works in two steps. The province database of addresses is
first downloaded from the National Open Database of Addresses, with URLs to
download the zip files available in \code{\link[cc.data]{buildings_osm_ms_keys}}.
For the missing addresses, we use the province's reverse geolocating service
(at the moment only available for QC and BC). Third, we reverse geocode using
the OSM service. The reverse geocoding is done using the centroid of every
spatial feature.
}
