% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{road_network_url}
\alias{road_network_url}
\title{URLs of where to get the road network shapefile}
\format{
\subsection{\code{road_network_url}}{

tibble with census years and URLs
}
}
\usage{
road_network_url
}
\description{
URLs of where to get the road network shapefile
}
\keyword{datasets}
