% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time_matrix.R
\name{tt_calculate_all_modes}
\alias{tt_calculate_all_modes}
\title{Calculate the travel time matrix of all mode}
\usage{
tt_calculate_all_modes(
  dest_folder,
  DA_table,
  routing_server = "http://localhost:5001/",
  osm_pbf = "north-america/canada-220101.osm.pbf"
)
}
\arguments{
\item{dest_folder}{<\code{character}> Destination folder where to download
and process the necessary data for the travel time matrix calculations.}

\item{DA_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame}

\item{routing_server}{<\code{character}> The base URL of the routing server. If
a server has been initiated using \code{\link[cc.data]{tt_local_osrm}},
the default URL will be \verb{http://localhost:5000/}. Other services can be used
like \verb{http://router.project-osrm.org/}.}

\item{osm_pbf}{<\code{character}> Link to the wanted version of OSM. Defaults to
Canada.}
}
\value{
Returns a list of three each holding a data.frame where the first
column is the DA ID, and all others are all IDs. Every cell is the trip duration
from and to DAs in seconds.
}
\description{
Calculate the travel time matrix of all mode
}
