% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Helper for interpolation}
\usage{
weighted_mean(x, w, ..., na.rm = TRUE)
}
\arguments{
\item{x}{The value}

\item{w}{The weight}

\item{...}{Other options to pass to stats::weighted.mean()}

\item{na.rm}{Remove NAs, defaults to TRUE}
}
\value{
Same output as weighted.mean, but with NA handled for weights.
}
\description{
Needed as stats::weighted.mean does not handle NA weights
}
