% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\name{zoning_collect_results}
\alias{zoning_collect_results}
\title{Collect mirai_map results with the native green ETA progress bar}
\usage{
zoning_collect_results(mp)
}
\arguments{
\item{mp}{A mirai_map promise object.}
}
\value{
The list of worker results.
}
\description{
Collect mirai_map results with the native green ETA progress bar
}
\keyword{internal}
