% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\docType{data}
\name{ZONING_USAGE_TYPES}
\alias{ZONING_USAGE_TYPES}
\title{Default residential usage types}
\format{
An object of class \code{character} of length 12.
}
\usage{
ZONING_USAGE_TYPES
}
\description{
Character vector of usage classes used by default across this module's
functions. Override via each function's \code{usage_types} argument.
}
\keyword{datasets}
