% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{bucket_get_folder}
\alias{bucket_get_folder}
\title{Read a bucket from AWS}
\usage{
bucket_get_folder(destination_folder, bucket, exclude = NULL)
}
\arguments{
\item{destination_folder}{<\code{character}> The folder in which the bucket should
be downloaded.}

\item{bucket}{<\code{character}> The name of the bucket from which to download
the objects.}

\item{exclude}{<\code{character}> Files to not download from the bucket. Defaults to
NULL for none (download everything).}
}
\value{
An error message or nothing if ran succesfully.
}
\description{
Download a whole amazon bucket in a destination folder.
}
