% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_custom_boundaries.R
\name{census_custom_boundaries}
\alias{census_custom_boundaries}
\title{Interpolate and process data for non-census boundaries}
\usage{
census_custom_boundaries(
  destination,
  data_raw_DA,
  census_vectors = susdata::census_vectors,
  census_scales = susdata::census_scales,
  census_years = susdata::census_years,
  agg_type = census_agg_type(census_vectors = census_vectors, census_scales =
    census_scales, census_years = census_years)
)
}
\arguments{
\item{destination}{<\verb{named list}> A named list of sf data.frame. The name is
the name of a scale which can be found in the census, e.g. \code{CSD}, \code{CT}, \code{DA}.
Each contains an sf data.frame of non-census boundaries, for which data must
be interpolated. e.g. boroughs in Montreal.}

\item{data_raw_DA}{<\verb{named list}> A named list of sf data.frame where the names
are census years. It must be the lower possible scale of the raw data (with parents),
and so the output of \code{\link[susdata]{census_data_raw}} \code{[["DA"]]}}

\item{census_vectors}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_vectors}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[susdata]{census_years}}}

\item{agg_type}{<\verb{named list}> The output of
\code{\link[susdata]{census_agg_type}}}
}
\value{
The same list as what is fed to the \code{destination} argument,
with all processed census data appended.
}
\description{
Interpolate and process data for non-census boundaries
}
