% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_raw.R
\name{census_data_raw}
\alias{census_data_raw}
\title{Get census vectors values}
\usage{
census_data_raw(
  empty_geometries,
  census_vectors_table = cc.data::census_vectors_table,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years
)
}
\arguments{
\item{empty_geometries}{<\verb{list of sf data.frame}> The output of
\code{\link[cc.data]{census_empty_geometries}}.}

\item{census_vectors_table}{<\code{data.frame}> Should be equal to
\code{\link[cc.data]{census_vectors_table}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[cc.data]{census_years}}}
}
\value{
A list of scales and years of census data accompanied with the
parent values.
}
\description{
Get census vectors values
}
