% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_raw.R
\name{census_data_raw}
\alias{census_data_raw}
\title{Get census vectors values}
\usage{
census_data_raw(
  census_empty_geometries,
  census_vectors,
  census_scales,
  census_years
)
}
\arguments{
\item{census_empty_geometries}{<\verb{list of sf data.frame}> The output of
\code{\link[susdata]{census_empty_geometries}}.}

\item{census_vectors}{<\code{data.frame}> Should be equal to
\code{\link[susdata]{census_vectors}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[susdata]{census_years}}}
}
\value{
A list of scales and years of census data accompanied with the
parent values.
}
\description{
Get census vectors values
}
