% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_drop_parents.R
\name{census_drop_parents}
\alias{census_drop_parents}
\title{Drop parent variables}
\usage{
census_drop_parents(normalized, census_vectors = cc.data::census_vectors)
}
\arguments{
\item{normalized}{<\verb{list of sf data.frame}> The output of
\code{\link[cc.data]{census_normalize}}.}

\item{census_vectors}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_vectors}}}
}
\value{
A list of scales and years of census data with the columns
corresponding to ID, census_vectors$var_code, and geometry.
}
\description{
Drop parent variables
}
