% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_empty_geometries.R
\name{census_empty_geometries}
\alias{census_empty_geometries}
\title{Get empty geometries}
\usage{
census_empty_geometries(census_years, census_scales)
}
\arguments{
\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[susdata]{census_years}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_scales}}}
}
\value{
A named list of \verb{sf data.frame} the same length as census_scales,
with census geometries, containing only ID, population and households.
}
\description{
Get empty geometries
}
