% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_interpolate.R
\name{census_interpolate}
\alias{census_interpolate}
\title{Interpolate all census years to the current year's geography}
\usage{
census_interpolate(data_raw, agg_type, census_scales, census_years)
}
\arguments{
\item{data_raw}{<\verb{list of sf data.frame}> The output of
\code{\link[susdata]{census_data_raw}}.}

\item{agg_type}{<\verb{named list}> The output of
\code{\link[susdata]{census_agg_type}}.}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[susdata]{census_years}}}
}
\value{
A list of scales and years of census data with geometries interpolated
to the current year.
}
\description{
Interpolate all census years to the current year's geography
}
