% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_interpolate.R
\name{census_interpolate}
\alias{census_interpolate}
\title{Interpolate all census years to the current year's geography}
\usage{
census_interpolate(
  data_raw,
  census_vectors_table = cc.data::census_vectors_table,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years,
  agg_type = census_agg_type(census_vectors_table = census_vectors_table, census_scales =
    census_scales, census_years = census_years)
)
}
\arguments{
\item{data_raw}{<\verb{list of sf data.frame}> The output of
\code{\link[cc.data]{census_data_raw}}.}

\item{census_vectors_table}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_vectors_table}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[cc.data]{census_years}}}

\item{agg_type}{<\verb{named list}> The output of
\code{\link[cc.data]{census_agg_type}}.}
}
\value{
A list of scales and years of census data with geometries interpolated
to the current year.
}
\description{
Interpolate all census years to the current year's geography
}
