% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_normalize.R
\name{census_normalize}
\alias{census_normalize}
\title{Normalize census variables to proportions in \verb{[0,1]}}
\usage{
census_normalize(
  interpolated,
  census_vectors = cc.data::census_vectors,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years,
  unit_type = NULL
)
}
\arguments{
\item{interpolated}{list [\link{scale}][\link{year}] sf — output of census_interpolate()}

\item{census_vectors}{\if{html}{\out{<character>}} cc.data::census_vectors (used only if unit_type = NULL)}

\item{census_scales}{\if{html}{\out{<character>}} cc.data::census_scales}

\item{census_years}{\if{html}{\out{<numeric>}}   cc.data::census_years}

\item{unit_type}{<data.frame|NULL> if NULL, built from variables actually present}
}
\value{
list [\link{scale}][\link{year}] sf with normalized variables
}
\description{
Normalize census variables to proportions in \verb{[0,1]}
}
