% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_normalize.R
\name{census_normalize}
\alias{census_normalize}
\title{Normalize percentage variables}
\usage{
census_normalize(
  interpolated,
  census_vectors,
  unit_type,
  census_scales,
  census_years
)
}
\arguments{
\item{interpolated}{<\verb{list of sf data.frame}> The output of
\code{\link[susdata]{census_interpolate}}.}

\item{census_vectors}{<\code{data.frame}> Should be equal to
\code{\link[susdata]{census_vectors}}}

\item{unit_type}{<\code{data.frame}> The output of
\code{\link[susdata]{census_unit_type}}.}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[susdata]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[susdata]{census_years}}}
}
\value{
A list of scales and years of census data with variables desired as
percentage normalized.
}
\description{
Normalize percentage variables
}
