% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_reduce_years.R
\name{census_reduce_years}
\alias{census_reduce_years}
\title{Reduce all census years to one dataframe}
\usage{
census_reduce_years(
  parent_dropped,
  census_vectors_table = cc.data::census_vectors_table,
  census_scales = cc.data::census_scales,
  census_years = cc.data::census_years
)
}
\arguments{
\item{parent_dropped}{<\verb{list of sf data.frame}> The output of
\code{\link[cc.data]{census_drop_parents}}.}

\item{census_vectors_table}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_vectors_table}}}

\item{census_scales}{<\verb{character vector}> Should be equal to
\code{\link[cc.data]{census_scales}}}

\item{census_years}{<\verb{numeric vector}> Should be equal to
\code{\link[cc.data]{census_years}}}
}
\value{
A list of scales with years reduced, and the years appended to all
variables column name
}
\description{
Reduce all census years to one dataframe
}
