% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_data.R
\name{cmhc_get_annual_csd}
\alias{cmhc_get_annual_csd}
\title{Downloads and reshapes annual CMHC Census Subdivision (CSD) data
for each year and CSD specified in the request list.
Data is harmonized to the 2021 geography using correspondence tables.}
\usage{
cmhc_get_annual_csd(requests, csd_correspondence_list, csd_uids)
}
\arguments{
\item{requests}{A list of request objects. Each must include \code{survey}, \code{series}, \code{dimension}, and \code{years}.}

\item{csd_correspondence_list}{A named list of correspondence tables to convert older CSD GeoUIDs to 2021 equivalents.}

\item{csd_uids}{A character vector of CSD GeoUIDs to include.}
}
\value{
A named list \code{CSD} containing reshaped and harmonized data frames for each variable.
}
\description{
Downloads and reshapes annual CMHC Census Subdivision (CSD) data
for each year and CSD specified in the request list.
Data is harmonized to the 2021 geography using correspondence tables.
}
