% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_create_user}
\alias{db_create_user}
\title{Create a read only user to the AWS MySQL database}
\usage{
db_create_user(user, password, admin = FALSE)
}
\arguments{
\item{user}{<\code{character}> User identifier for the new user}

\item{password}{<\code{character}> Password identifier for the new user}

\item{admin}{<\code{logical}> If the user created should be granted all powers
on the MySQL db. If FALSE, read only permissions are attributed.}
}
\value{
Returns an error or nothing if ran successfully. User is then created and
has SELECT and SHOW VIEW privileges only if \code{admin} is set to FALSE (default).
}
\description{
Create a read only user to the AWS MySQL database
}
