% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_create_user}
\alias{db_create_user}
\title{Create a read only user to the AWS MySQL database}
\usage{
db_create_user(conn, user, password, admin = FALSE)
}
\arguments{
\item{conn}{A \code{\link[DBI]{DBIConnection-class}} object, as returned by
\code{\link[cc.data]{db_connect}}.}

\item{user}{<\code{character}> User identifier for the new user}

\item{password}{<\code{character}> Password identifier for the new user}

\item{admin}{<\code{logical}> If the user created should be granted all powers
on the MySQL db. If FALSE, read only permissions are attributed.}
}
\value{
Returns an error or a success message. User is then created and
has SELECT and SHOW VIEW privileges only.
}
\description{
Create a read only user to the AWS MySQL database
}
