% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_delete_user}
\alias{db_delete_user}
\title{Delete a user that has access to the AWS MySQL database}
\usage{
db_delete_user(user)
}
\arguments{
\item{user}{<\code{character}> User identifier for the user to remove}
}
\value{
Returns an error or nothing if ran successfully. User's access are
then removed and the user is deleted.
}
\description{
Delete a user that has access to the AWS MySQL database
}
