% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_query}
\alias{db_query}
\title{Execute, write or get queries from the MySQL database}
\usage{
db_query(type, statement, name = NULL)
}
\arguments{
\item{type}{<\code{character}> One of \code{execute}, \code{write} or \code{get}. They correspond
to \code{DBI::dbExecute}, \code{DBI::dbWriteTable} and \code{DBI::dbGetQuery}.}

\item{statement}{<\code{character} or \code{data.frame}> A character string containing SQL
for type \code{execute} and \code{get}, but a data.frame for type \code{write}.}

\item{name}{<\code{character}> The table name, mandatory for type \code{write}. In other
cases, won't be used.}
}
\value{
An error or nothing if ran successfully.
}
\description{
Execute, write or get queries with connection and disconnection included, to
ensure that connections are as short-lived as possible, and that no
connection remains open indefinitely.
}
