% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{db_write_processed_data}
\alias{db_write_processed_data}
\title{Write the list of processed census data to the MySQL database}
\usage{
db_write_processed_data(processed_census)
}
\arguments{
\item{processed_census}{<\verb{named list}> The finalized process data
normally coming out of the final processing function:
\code{\link[cc.data]{census_reduce_years}}}
}
\value{
Returns an error or nothing if ran successfully. All tables in the
list fed to \code{processed_census} are written to the MySQL database with
the spatial features dropped.
}
\description{
Write the list of processed census data to the MySQL database
}
