% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_data.R
\name{format_date_names}
\alias{format_date_names}
\title{Helper: Reformat "YYYY-MM" to "YYYYMM"}
\usage{
format_date_names(x)
}
\arguments{
\item{x}{\if{html}{\out{<character>}} Vector of strings (e.g., column names).}
}
\value{
\if{html}{\out{<character>}} The input \code{x} with hyphens removed.
}
\description{
Remove hyphens from strings that look like year-month ("YYYY-MM") so
they become "YYYYMM".
}
