% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foursquare_poi.R
\name{foursquare_connect}
\alias{foursquare_connect}
\title{Connect to Foursquare Database}
\usage{
foursquare_connect()
}
\value{
A database connection object.
}
\description{
Establishes a connection to the Snowflake Foursquare database using environment variables.
}
