% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode.R
\name{geocode_localhost}
\alias{geocode_localhost}
\title{Geocode using a local Nominatim instance (docker)}
\usage{
geocode_localhost(address)
}
\arguments{
\item{address}{< \code{character}> Character vector corresponding to an address,
e.g. 6540, Avenue Baldwin, Anjou, Montreal}
}
\value{
\code{sf} point
}
\description{
Geocode using a local Nominatim instance (docker)
}
