% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canale.R
\name{get_all_three_plus_ways}
\alias{get_all_three_plus_ways}
\title{Get all three+ way intersections in Canadian streets for a given year}
\usage{
get_all_three_plus_ways(year)
}
\arguments{
\item{year}{<\code{numeric}> indicates the year of the streets network to use.
Options are census years starting 2001.}
}
\value{
An \code{sf} object representing the three+ way intersections in the
streets network.
}
\description{
This function downloads the streets network for a given year and returns all
the intersections with three or more streets.
}
