% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foursquare_poi.R
\name{poi_run_evolution}
\alias{poi_run_evolution}
\title{Process POI Evolution}
\usage{
poi_run_evolution(geo_level, census_year)
}
\arguments{
\item{geo_level}{A string specifying the census level: "C", "PR", "CMA", "CSD", "CT", or "DA".}

\item{census_year}{A string specifying the census dataset code (default = "CA21").}
}
\value{
An \code{sf} object with columns: \code{id}, yearly POI totals.
}
\description{
Connects to Foursquare, retrieves POIs, downloads the selected census geometry, intersects both, and returns a wide-format dataset of cumulative POIs by year and region.
}
