% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets_sf.R
\name{streets_sf}
\alias{streets_sf}
\title{Build streets country-wide}
\usage{
streets_sf(DA_processed_table, nb_batches = 500)
}
\arguments{
\item{DA_processed_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame from
which the DA ID will be attached to every single building.}

\item{nb_batches}{<\code{numeric}> In how many batches should the sf that
are to be reverse geocoded through the local instance of Nominatim be divided?
To be gentle on memory.}
}
\value{
A street segment sf dataframe reverse geocoded.
}
\description{
This function uses the \code{\link[cc.data]{rev_geocode_localhost}} function,
which means a local version of Nominatim must have been already set up and
needs to be running on port 8080. You can do so using
\code{\link[cc.data]{rev_geocode_create_local_nominatim}}.
}
