% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time_matrix.R
\name{tt_calculate}
\alias{tt_calculate}
\title{Create a travel time matrix using Dissemination Areas or Dissemination Blocks}
\usage{
tt_calculate(
  centroids,
  max_dist = 120000,
  routing_server = "http://localhost:5001/"
)
}
\arguments{
\item{centroids}{<\verb{sf data.frame}> Points (DA ou DB) servant d'origines
et de destinations.}

\item{max_dist}{<\code{numeric}> Distance max utilisée pour couper les destinations.
Exemples: \code{car} ~ 120000 (120 km/h), \code{bicycle} ~ 30000 (30 km/h),
\code{foot} ~ 10000 (10 km/h).}

\item{routing_server}{<\code{character}> Base URL du serveur de routage.
Si \code{\link{tt_local_osrm}} a été lancé localement, la valeur par défaut
est \verb{http://localhost:5001/}. D'autres services peuvent être utilisés,
p.ex. \verb{http://router.project-osrm.org/}.}
}
\value{
The travel time matrix
}
\description{
Create a travel time matrix using Dissemination Areas or Dissemination Blocks
}
