% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time_matrix.R
\name{tt_local_osrm}
\alias{tt_local_osrm}
\title{Create a docker local OSRM instance for faster travel time matrix}
\usage{
tt_local_osrm(
  dest_folder,
  mode = "car",
  osm_pbf = "http://download.geofabrik.de/north-america/canada-220101.osm.pbf"
)
}
\arguments{
\item{dest_folder}{<\code{character}> Destination folder where to download
and process the necessary data for the travel time matrix calculations.}

\item{mode}{<\code{character}> Options are \code{bicycle}, \code{car} and \code{foot}. Here
are the profiles:
\href{https://github.com/Project-OSRM/osrm-backend/tree/master/profiles}{OSRM profiles}}

\item{osm_pbf}{<\code{character}> Link to the wanted version of OSM. Defaults to
Canada.}
}
\value{
Opens a terminal from which the docker image is created.
}
\description{
For further information, read the
\href{https://github.com/Project-OSRM/osrm-backend}{Project-OSRM/osrm-backend}
documentation
}
