% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time_matrix.R
\name{tt_local_osrm}
\alias{tt_local_osrm}
\title{Create a local OSRM instance with Docker}
\usage{
tt_local_osrm(
  mode = "car",
  port = 5001L,
  osm_pbf = "north-america/canada-latest.osm.pbf",
  dest_folder = tempdir()
)
}
\arguments{
\item{mode}{\if{html}{\out{<character>}} One of \code{"car"}, \code{"bicycle"}, or \code{"foot"}. Default \code{"car"}.}

\item{port}{\if{html}{\out{<integer>}} Port for the OSRM HTTP server. Default \code{5001L}.}

\item{osm_pbf}{\if{html}{\out{<character>}} Path or relative name of the OSM \code{.pbf} to use.
Default \code{"north-america/canada-latest.osm.pbf"}.}

\item{dest_folder}{\if{html}{\out{<character>}} Directory where data will be downloaded and
processed. Default \code{tempdir()}.}
}
\value{
\if{html}{\out{<character>}} (invisibly) the \code{dest_folder}. Function is called for its side effects.
}
\description{
Downloads OSM data (if needed), builds the OSRM graph for the chosen \code{mode},
and launches a local OSRM HTTP server in a Docker container.
For details, see
\href{https://github.com/Project-OSRM/osrm-backend}{Project-OSRM/osrm-backend}
and
\href{https://github.com/Project-OSRM/osrm-backend/tree/master/profiles}{OSRM profiles}.
}
