% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\docType{data}
\name{ZONING_USAGE_TYPES}
\alias{ZONING_USAGE_TYPES}
\title{Residential usage types carried by the zoning dataset}
\format{
An object of class \code{character} of length 12.
}
\usage{
ZONING_USAGE_TYPES
}
\description{
The 12 residential usage categories encoded in the \code{usage_classification}
list-column of the zoning GeoJSON. Each zone permits one or more of these.
}
\keyword{internal}
